requirements.txt
fastapi==0.115.14
uvicorn==0.29.0
python-dotenv==1.0.1
openai==0.28.1
easyocr==1.7.2
pydantic==2.7.1
shutilwhich==1.1.0
Pillow==10.3.0
torch==2.2.2
numpy==1.26.4
typing-extensions==4.11.0
python-multipart